More error detection in TicketMeta.__init__
ID and title are now required.  Reorganized constructor to check for errors;
it is now called from `from_lines` instead of taking lines directly.
Status and creation time are defaulted.

Convert-from-string is now a class nethod

New TicketStatus class that can only be one of the expected statuses

More tests